Fix imconfig issues

+++ src/main/MQImguiWidgets.cpp
@@ -589,7 +589,8 @@ void DrawEQText(ImDrawList* drawList, int fontStyle, const char* text, const CXR
 		{
 			for (auto& charPos : textLine.arCharPos)
 			{
-				eqFont->RenderChar(drawList, eqFont->FontSize, ImVec2(textLine.nXOffset + charPos.x, textLine.y + yOffset),
+				eqFont->RenderChar(drawList, eqFont->FontSize,
+					ImVec2(static_cast<float>(textLine.nXOffset + charPos.x), static_cast<float>(textLine.y + yOffset)),
 					color.ToImU32(), charPos.c);
 			}
 		}
